Refactor codebase with Feature Sliced Design and colocation
Implement Feature Sliced Design (FSD) architecture with proper layer separation and colocation principles for improved maintainability and scalability.

## Major Changes

### FSD Layer Structure
- **app/**: Application entry point, configuration, and providers
- **entities/**: Domain entities with colocation (api, model, lib, __tests__)
- **features/**: Individual feature slices for each operation
- **shared/**: Shared utilities, database, and error handling

### Colocation Implementation
- Moved embedding-related tests to `src/entities/embedding/__tests__/`
- Grouped related API, model, and lib code within entities
- Created feature slices for each embedding operation:
  - create-embedding
  - batch-create-embedding
  - search-embeddings
  - list-embeddings
  - delete-embedding

### File Reorganization
- `src/services/` → `src/entities/embedding/api/`
- `src/types/` → `src/entities/embedding/model/`
- `src/schemas/` → `src/entities/embedding/model/`
- `src/utils/` → `src/entities/embedding/lib/` + `src/shared/lib/`
- `src/database/` → `src/shared/database/`
- `src/errors/` → `src/shared/errors/`

### Updated Import Paths
- Fixed all import statements to reflect new FSD structure
- Updated test files with correct module paths
- Maintained TypeScript type safety throughout refactoring

### Technical Improvements
- Environment variable access updated for TypeScript strict mode
- Replaced unsafe `any` types with proper type annotations in tests
- Maintained full application functionality
- Preserved Effect-TS functional programming patterns
- Build process remains fully functional
