+++ src/__tests__/database.test.ts
@@ -3,8 +3,11 @@ import { createClient } from "@libsql/client"
 import { drizzle } from "drizzle-orm/libsql"
 import { Effect, Exit } from "effect"
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest"
-import { DatabaseService, DatabaseServiceLive } from "../database/connection"
-import { DatabaseConnectionError } from "../errors/database"
+import {
+  DatabaseService,
+  DatabaseServiceLive,
+} from "../shared/database/connection"
+import { DatabaseConnectionError } from "../shared/errors/database"
 
 // Mock fs module for directory creation tests
 vi.mock("fs", () => ({
