+++ src/entities/embedding/api/index.ts
@@ -1,0 +1,3 @@
+// Embedding API exports
+export * from "./embedding"
+export * from "./ollama"
