+++ src/entities/embedding/index.ts
@@ -1,0 +1,5 @@
+// Embedding entity public API
+
+export * from "./api"
+export * from "./lib"
+export * from "./model"
