+++ src/entities/embedding/lib/embedding-data.ts
@@ -2,7 +2,7 @@ import { Effect } from "effect"
 import {
   EmbeddingVectorSchema,
   StoredEmbeddingDataSchema,
-} from "../schemas/openapi"
+} from "../model/openapi"
 
 export class EmbeddingDataParseError extends Error {
   constructor(message: string, cause?: unknown) {
