+++ src/entities/embedding/lib/index.ts
@@ -1,0 +1,2 @@
+// Embedding lib exports
+export * from "./embedding-data"
