+++ src/entities/embedding/model/index.ts
@@ -1,0 +1,3 @@
+// Embedding model exports
+export * from "./embedding"
+export * from "./openapi"
