+++ src/entities/embedding/model/openapi.ts
