+++ src/features/batch-create-embedding/index.ts
@@ -1,0 +1,1 @@
+export { batchCreateEmbeddingRoute } from "./api/route"
