+++ src/features/create-embedding/index.ts
@@ -1,0 +1,1 @@
+export { createEmbeddingRoute } from "./api/route"
