+++ src/features/delete-embedding/index.ts
@@ -1,0 +1,1 @@
+export { deleteEmbeddingRoute } from "./api/route"
