+++ src/features/list-embeddings/index.ts
@@ -1,0 +1,1 @@
+export { getEmbeddingByUriRoute, listEmbeddingsRoute } from "./api/route"
