+++ src/features/search-embeddings/index.ts
@@ -1,0 +1,1 @@
+export { searchEmbeddingsRoute } from "./api/route"
