+++ src/shared/errors/database.ts
