+++ src/shared/errors/index.ts
@@ -1,0 +1,2 @@
+export * from "./database"
+export * from "./ollama"
