+++ src/shared/errors/ollama.ts
